+++ src/main/java/com/atlauncher/gui/panels/packbrowser/FTBPacksPanel.java
@@ -166,6 +166,13 @@ public class FTBPacksPanel extends PackBrowserPlatformPanel {
         return true;
     }
 
+    public boolean supportsManualAdding() {
+        return false;
+    }
+
+    public void addById(String id) {
+    }
+
     @Override
     public String getPlatformMessage() {
         return ConfigManager.getConfigItem("platforms.modpacksch.message", null);