+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This changelog only contains the changes that are unreleased. For changes for in
 - Use image for login to Microsoft button
 - Add ability to change sort order for ATLauncher and CurseForge packs
 - Add in Name and Featured sort fields for CurseForge [#554]
+- Add ability to add a CurseForge/Modrinth pack from the packs browser by id/slug/url [#554]
 
 ### Fixes
 - Remove Vercel links from being valid from Modrinth imports